montailleur/python-ai/app/routers/virtual_tryon.py
"""
Virtual Try-On Router
Handles AI-powered virtual clothing try-on
"""

from fastapi import APIRouter, File, UploadFile, HTTPException, Form
from pydantic import BaseModel
from typing import Optional
import io
from PIL import Image

router = APIRouter()


class VirtualTryOnRequest(BaseModel):
    garment_type: str  # dress, shirt, pants, etc.
    garment_color: Optional[str] = None


class VirtualTryOnResponse(BaseModel):
    success: bool
    result_url: Optional[str] = None
    status: str
    message: str


@router.post("/try-on", response_model=VirtualTryOnResponse)
async def virtual_try_on(
    person_image: UploadFile = File(...),
    garment_image: UploadFile = File(...),
    garment_type: str = Form(...)
):
    """
    Generate a virtual try-on image.
    
    - **person_image**: Photo of the person
    - **garment_image**: Photo of the garment
    - **garment_type**: Type of garment (dress, shirt, pants, etc.)
    
    Returns the processed image with the garment virtually fitted.
    """
    try:
        # Read images
        person_contents = await person_image.read()
        garment_contents = await garment_image.read()
        
        person_img = Image.open(io.BytesIO(person_contents))
        garment_img = Image.open(io.BytesIO(garment_contents))
        
        # TODO: Implement actual virtual try-on AI model
        # For now, return a placeholder response
        
        return VirtualTryOnResponse(
            success=True,
            result_url="https://example.com/result.jpg",  # Placeholder
            status="completed",
            message="Virtual try-on completed successfully"
        )
        
    except Exception as e:
        raise HTTPException(status_code=400, detail=str(e))


@router.post("/try-on-status/{job_id}")
async def get_try_on_status(job_id: str):
    """
    Get the status of a virtual try-on job.
    """
    # TODO: Implement job tracking
    return {
        "job_id": job_id,
        "status": "completed",
        "progress": 100,
        "result_url": "https://example.com/result.jpg"
    }


@router.get("/garment-types")
async def get_garment_types():
    """
    Get list of supported garment types.
    """
    return {
        "garment_types": [
            {"id": "dress", "name": "Robe", "icon": "dress"},
            {"id": "shirt", "name": "Chemise", "icon": "shirt"},
            {"id": "pants", "name": "Pantalon", "icon": "pants"},
            {"id": "skirt", "name": "Jupe", "icon": "skirt"},
            {"id": "jacket", "name": "Veste", "icon": "jacket"},
            {"id": "suit", "name": "Costume", "icon": "suit"},
            {"id": "traditional", "name": "Tenue traditionnelle", "icon": "traditional"},
        ]
    }
