montailleur/python-ai/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image Processing
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.3

# Machine Learning
mediapipe==0.10.9
tensorflow==2.15.0
scikit-learn==1.4.0

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
